--- store/rabbitmq/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "graph-store-rabbitmq"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-amiquip = "0.4.2"
-async-trait = "0.1.58"
-graph = { path = "../../graph" }
